+++ packages/dashboard/src/components/Card/ActionBar/ActionBar.tsx
@@ -93,7 +93,8 @@ export class ActionBar extends React.Component<IActionBarProps, {}> {
           calloutProps: {
             styles: {
               root: {
-                boxShadow: '0px 1.2px 3.6px rgba(0,0,0,0.18), 0px 6.4px 14.4px rgba(0,0,0,0.22)'
+                boxShadow: '0px 1.2px 3.6px rgba(0,0,0,0.18), 0px 6.4px 14.4px rgba(0,0,0,0.22)',
+                border: 'none'
               }
             }
           },