+++ packages/date-time/src/components/Calendar/CalendarMonth/CalendarMonth.tsx
@@ -1,0 +1,13 @@
+import { ICalendarMonthProps, ICalendarMonthStyleProps, ICalendarMonthStyles } from './CalendarMonth.types';
+import { CalendarMonthBase } from './CalendarMonth.base';
+import { getStyles } from './CalendarMonth.styles';
+import { styled } from 'office-ui-fabric-react/lib/Utilities';
+
+/**
+ * CalendarMonth description
+ */
+export const CalendarMonth: (props: ICalendarMonthProps) => JSX.Element = styled<
+  ICalendarMonthProps,
+  ICalendarMonthStyleProps,
+  ICalendarMonthStyles
+>(CalendarMonthBase, getStyles, undefined, { scope: 'CalendarMonth' });