+++ packages/office-ui-fabric-react/src/components/ComboBox/VirtualizedComboBox.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { BaseComponent, createRef } from '../../Utilities';
+import { BaseComponent } from '../../Utilities';
 import { ComboBox } from './ComboBox';
 import { IComboBoxProps, IComboBox } from './ComboBox.types';
 import { IList, List } from '../../List';
@@ -7,9 +7,9 @@ import { ISelectableOption } from '../../utilities/selectableOption/SelectableOp
 
 export class VirtualizedComboBox extends BaseComponent<IComboBoxProps, {}> implements IComboBox {
   /** The combo box element */
-  private _comboBox = createRef<IComboBox>();
+  private _comboBox = React.createRef<IComboBox>();
   /** The virtualized list element */
-  private _list = createRef<IList>();
+  private _list = React.createRef<IList>();
 
   public dismissMenu(): void {
     if (this._comboBox.current) {