+++ packages/office-ui-fabric-react/src/components/ContextualMenu/ContextualMenu.base.tsx
@@ -206,6 +206,7 @@ export class ContextualMenuBase extends BaseComponent<IContextualMenuProps, ICon
       beakWidth,
       directionalHint,
       directionalHintForRTL,
+      alignTargetEdge,
       gapSpace,
       coverTarget,
       ariaLabel,
@@ -309,6 +310,7 @@ export class ContextualMenuBase extends BaseComponent<IContextualMenuProps, ICon
           onScroll={this._onScroll}
           bounds={bounds}
           directionalHintFixed={directionalHintFixed}
+          alignTargetEdge={alignTargetEdge}
           hidden={this.props.hidden}
         >
           <div
@@ -457,10 +459,15 @@ export class ContextualMenuBase extends BaseComponent<IContextualMenuProps, ICon
         primaryDisabled: item.primaryDisabled
       };
 
+      const menuItemStyles = this._classNames.subComponentStyles
+        ? (this._classNames.subComponentStyles.menuItem as IStyleFunctionOrObject<IContextualMenuItemStyleProps, IContextualMenuItemStyles>)
+        : undefined;
+
       // We need to generate default styles then override if styles are provided
       // since the ContextualMenu currently handles item classNames.
       itemClassNames = mergeStyleSets(
         getContextualMenuItemClassNames(getItemStyles, itemStyleProps),
+        getContextualMenuItemClassNames(menuItemStyles, itemStyleProps),
         getContextualMenuItemClassNames(styles, itemStyleProps)
       );
     }